Revert "Merge branch 'FollowingPokemon' into master"
This reverts commit 435a4f84e24d7cf678610cfe18810a3cb8370847, reversing
changes made to bee11f31cd460d5df86129c31bf7d91143a96e6e.

+++ src/fldeff_rocksmash.c
@@ -27,8 +27,7 @@ static void Task_DoFieldMove_RunFunc(u8 taskId);
 static void FieldCallback_RockSmash(void);
 static void FieldMove_RockSmash(void);
 
-// text
-bool8 CheckObjectGraphicsInFrontOfPlayer(u16 graphicsId)
+bool8 CheckObjectGraphicsInFrontOfPlayer(u8 graphicsId)
 {
     u8 objEventId;
